requirements.txt
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
DateTime==5.1
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
exceptiongroup==1.1.1
fastjsonschema==2.16.3
fonttools==4.39.3
fqdn==1.5.1
frozendict==2.3.7
html5lib==1.1
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.0
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.6.3
jupyter-core==4.8.1
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterthemes==0.20.0
kiwisolver==1.4.4
lesscpy==0.15.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.3
mistune==2.0.5
multitasking==0.0.11
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.1
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.2
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pid==3.0.4
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.20
psutil==5.9.5
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.4.2
QtPy==2.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.1
traitlets==5.1.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
yfinance==0.2.18
zipp==3.15.0
zope.interface==6.0

python-dotenv~=1.0.0
sklearn~=0.0.post5
scikit-learn~=1.2.2